Deleted groovy html files, some refactoring around engine and player.

+++ src/gameplayer/controller/PlayerController.java
@@ -3,5 +3,5 @@ package gameplayer.controller;
 public interface PlayerController {
     void levelWon(int level);
 
-    void lifeChange(Double livesLeft);
+    void lifeChange(Integer livesLeft);
 }
